Fixed StringID bitmasks for Reach and H4 (fixes #52), also fixed H3 script objects loading incorrectly

+++ src/Blamite/Blam/ThirdGen/Resources/Models/ThirdGenModelPermutation.cs
@@ -20,7 +20,7 @@ namespace Blamite.Blam.ThirdGen.Resources.Models
 
         private void Load(StructureValueCollection values)
         {
-            Name = new StringID((int)values.GetInteger("name stringid"));
+            Name = new StringID(values.GetInteger("name stringid"));
             ModelSectionIndex = (int)values.GetInteger("model section");
         }
     }
